Spreadsheet (#3)
* In progress spreadsheet file

* Update spreadsheet.py

* trying to access spreadsheet

* Accessing the google spreadsheet v1

* Made spreadsheet fetching a function

* Spreadsheet fetching as a function

* fix

* Connecting the form to the selection_routes

We connected to the website the form html life that we did.

* vf fetching spreadsheet data

* Connect spreadsheet and routes

Connected the spreadsheet with the form and result routes. Must now format the results in the "Filtered" page to show the house and the google street view with it.

* Update of form

Reduce requirements of the filtering

* Updated Filtered data -html v1

First update of the data display from the form's results. Shows it in the shape of a container with blue edges

* Retired files not used

* v2 of display of data

Tried to get Street View into the filtered data display. Currently NOT working as we can't retrieve the correct Street View link

* Now street view works!!!! Yay

Need to fix the API key thing but making progress

---------

Co-authored-by: ariella9 <ai396@georgetown.edu>
Co-authored-by: vsaab1 <vs647@georgetown.edu>